+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+- jrl_release: add `ConanFileVersionExtractor` by @arntanguy
+- jrl_release: add `--check-version --check-tag vX.Y.Z` by @arntanguy
 - jrl_release: add `DebianChangelogVersionExtractor` by @arntanguy
 
 ## [2.1.0] - 2026-07-03